Restrict pages based on authentication

+++ base/views.py
@@ -1,12 +1,15 @@
 from urllib.request import Request
 from django.shortcuts import render, redirect
-# from django.http import HttpResponse
+from django.http import HttpResponse
 from .models import Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+
+# restricting pages
+from django.contrib.auth.decorators import login_required
 
 # rooms = [
 #     {"id":1, "name":"Learn JavaScript"},
@@ -21,6 +24,10 @@ queryset = ModelName.objects.all()
 """
 
 def login_page(request):
+    if request.user.is_authenticated:
+        return redirect("home")
+
+
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -72,7 +79,7 @@ def room(request, pk):
     context = {"room": room}
     return render(request, "base/room.html", context)
 
-
+@login_required(login_url="login") # restrict to logged in users only
 def create_room(request):
     form = RoomForm()
     if request.method == "POST": # post data
@@ -85,9 +92,13 @@ def create_room(request):
     context = {"form": form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url="login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You don't have sufficient permissions to view this.")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -99,8 +110,13 @@ def update_room(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse("You don't have sufficient permissions to view this.")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
